+++ dascore/proc/basic.py
@@ -10,21 +10,18 @@ import pandas as pd
 from scipy.fft import next_fast_len
 
 import dascore as dc
-from dascore.constants import DEFAULT_ATTRS_TO_IGNORE, PatchType
+from dascore.constants import PatchType
 from dascore.core.attrs import PatchAttrs
 from dascore.core.coordmanager import CoordManager, get_coord_manager
 from dascore.core.coords import get_coord
-from dascore.exceptions import ParameterError, PatchBroadcastError, UnitError
-from dascore.units import DimensionalityError, Quantity, Unit, get_quantity
+from dascore.exceptions import ParameterError
 from dascore.utils.misc import _get_nullish
 from dascore.utils.models import ArrayLike
 from dascore.utils.patch import (
-    _merge_aligned_coords,
-    _merge_models,
-    align_patch_coords,
     get_dim_axis_value,
     patch_function,
 )
+from dascore.utils.ufuncs import apply_ufunc
 
 
 def set_dims(self: PatchType, **kwargs: str) -> PatchType:
@@ -154,6 +151,17 @@ def equals(self: PatchType, other: Any, only_required_attrs=True, close=False) -
     elif not close and not np.equal(self.data, other.data).all():
         return False
     return True
+
+
+def bool_patch(self: PatchType):
+    """
+    Get the boolean value of a patch.
+
+    This follows the numpy convention of raising a value error if the patch
+    has more than on element. Otherwise, it returns the truthy value of the
+    one element.
+    """
+    return bool(self.data)
 
 
 def update(
@@ -361,131 +369,9 @@ def standardize(
     return self.new(data=new_data)
 
 
-@patch_function()
-def apply_ufunc(
-    patch: PatchType | ArrayLike,
-    other: PatchType | ArrayLike,
-    operator: Callable,
-    *args,
-    attrs_to_ignore=DEFAULT_ATTRS_TO_IGNORE,
-    **kwargs,
-) -> PatchType:
-    """
-    Apply a ufunc-type operator to a patch.
-
-    This is used to implement a patch's operator overloading.
-
-    Parameters
-    ----------
-    patch
-        The patch instance.
-    other
-        The other object to apply the operator element-wise. Must be either a
-        non-patch which is broadcastable to the shape of the patch's data, or
-        a patch which has compatible coordinates. If units are provided they
-        must be compatible.
-    operator
-        The operator. Must be numpy ufunc-like.
-    *args
-        Arguments to pass to the operator.
-    attrs_to_ignore
-        Attributes to ignore when considering if patches are compatible.
-    **kwargs
-        Keyword arguments to pass to the operator.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> import dascore as dc
-    >>> from dascore.proc.basic import apply_ufunc
-    >>> patch = dc.get_example_patch()
-    >>> # multiply the patch by 10
-    >>> new = apply_ufunc(patch, 10, np.multiply)
-    >>> assert np.allclose(patch.data * 10, new.data)
-    >>> # add a random value to each element of patch data
-    >>> noise = np.random.random(patch.shape)
-    >>> new = apply_ufunc(patch, noise, np.add)
-    >>> assert np.allclose(new.data, patch.data + noise)
-    >>> # subtract one patch from another. Coords and attrs must be compatible
-    >>> new = apply_ufunc(patch, patch, np.subtract)
-    >>> assert np.allclose(new.data, 0)
-
-    Notes
-    -----
-    See [numpy's ufunc docs](https://numpy.org/doc/stable/reference/ufuncs.html)
-    """
-
-    def _get_coords_attrs_from_patches(patch, other):
-        """Deal with aligning two patches."""
-        # Align patches so their coords are identical and data aligned.
-        patch, other_patch = align_patch_coords(patch, other)
-        coords = _merge_aligned_coords(patch.coords, other_patch.coords)
-        # Get new attributes.
-        attrs = _merge_models(
-            patch.attrs,
-            other_patch.attrs,
-            attrs_to_ignore=attrs_to_ignore,
-        )
-        other = other_patch.data
-        if other_units := get_quantity(other_patch.attrs.data_units):
-            other = other * other_units
-        return patch, other, coords, attrs
-
-    def _ensure_array_compatible(patch, other):
-        """Deal with broadcasting a patch and an array."""
-        # This handles warning from quantity.
-        other = other.magnitude if hasattr(other, "magnitude") else other
-        other = np.asanyarray(other)
-        if patch.shape == other.shape:
-            return patch
-        if (patch_ndims := patch.ndim) < (array_ndims := other.ndim):
-            msg = f"Cannot broadcast patch/array {patch_ndims=} {array_ndims=}"
-            raise PatchBroadcastError(msg)
-        patch = patch.make_broadcastable_to(other.shape)
-        return patch
-
-    def _apply_op(array1, array2, operator, reversed=False):
-        """Simply apply the operator, account for reversal."""
-        if reversed:
-            array1, array2 = array2, array1
-        return operator(array1, array2, *args, **kwargs)
-
-    def _apply_op_units(patch, other, operator, attrs, reversed=False):
-        """Apply the operation handling units attached to array."""
-        data_units = get_quantity(attrs.data_units)
-        data = patch.data if data_units is None else patch.data * data_units
-        # other is not numpy array wrapped w/ quantity, convert to quant
-        if not hasattr(other, "shape"):
-            other = get_quantity(other)
-        try:
-            new_data_w_units = _apply_op(data, other, operator, reversed=reversed)
-        except DimensionalityError:
-            msg = f"{operator} failed with units {data_units} and {other.units}"
-            raise UnitError(msg)
-        attrs = attrs.update(data_units=str(new_data_w_units.units))
-        new_data = new_data_w_units.magnitude
-        return new_data, attrs
-
-    reversed = False  # flag to indicate we need to reverse data and patch
-    if not isinstance(patch, dc.Patch):
-        reversed = True
-        patch, other = other, patch
-    # Align/broadcast patch to input
-    if isinstance(other, dc.Patch):
-        patch, other, coords, attrs = _get_coords_attrs_from_patches(patch, other)
-    else:
-        patch = _ensure_array_compatible(patch, other)
-        coords, attrs = patch.coords, patch.attrs
-    # Apply operation
-    if isinstance(other, Quantity | Unit):
-        new_data, attrs = _apply_op_units(patch, other, operator, attrs, reversed)
-    else:
-        new_data = _apply_op(patch.data, other, operator, reversed)
-    new = patch.new(data=new_data, coords=coords, attrs=attrs)
-    return new
-
-
-# This is left here to not break compatibility.
+# This is left here to not break compatibility. It also forces `apply_ufunc`
+# to be imported in to this module where it used to live. In the rare chance
+# a user tries to access it directly from here it will still work.
 apply_operator = apply_ufunc
